# Conflicts:
#   comp/src/gb.web.gui/.lang/it.po

Contains correction and new Italian translations

[DEVELOPMENT ENVIRONMENT]
* NEW: The IDE translation is now complete.

[WIKI]
* NEW: The wiki translation is now complete.

[GB.ARG]
* NEW: The arg translation is now complete.

[GB.FORM.MDI]
* NEW: The mdi translation is now complete.

[GB.HIGHLIGHT]
* NEW: The highlight translation is now complete.

+++ gb.gtk/src/gdrawingarea.cpp
@@ -399,7 +399,7 @@ void gDrawingArea::setBackground(gColor color)
 }
 #endif
 
-long gDrawingArea::handle()
+uintptr_t gDrawingArea::handle()
 {
 	if (!_own_window)
 	{
